Fix Object denormalizers
Denormalize objects using APIP denormalizers (except for HAL because APIP doesn't handle it)
