+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.1] - 2020-05-12
+* Denormalize objects using APIP denormalizers (except for HAL because APIP doesn't handle it)
 
 ## [0.1.0] - 2020-05-10
 * Added microservice specific HTTP client
