--- src/Serializer/AbstractObjectDenormalizer.php
@@ -1,42 +1,0 @@
-<?php
-
-namespace Mtarld\ApiPlatformMsBundle\Serializer;
-
-use Symfony\Component\Serializer\Normalizer\CacheableSupportsMethodInterface;
-use Symfony\Component\Serializer\Normalizer\DenormalizerAwareInterface;
-use Symfony\Component\Serializer\Normalizer\DenormalizerAwareTrait;
-use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
-
-/**
- * @internal
- * @SuppressWarnings(PHPMD.UnusedFormalParameter)
- */
-abstract class AbstractObjectDenormalizer implements DenormalizerInterface, DenormalizerAwareInterface, CacheableSupportsMethodInterface
-{
-    use DenormalizerAwareTrait;
-
-    abstract protected function getFormat(): string;
-
-    /**
-     * @param string $type
-     * @param string $format
-     */
-    public function denormalize($data, $type, $format = null, array $context = [])
-    {
-        return $this->denormalizer->denormalize($data, $type, 'json');
-    }
-
-    /**
-     * @param string $type
-     * @param string $format
-     */
-    public function supportsDenormalization($data, $type, $format = null): bool
-    {
-        return $this->getFormat() === $format;
-    }
-
-    public function hasCacheableSupportsMethod(): bool
-    {
-        return true;
-    }
-}
